+++ tests/Stubs/MyDateTime.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Somnambulist\EntityAudit\Tests\Stubs;
+
+use DateTimeImmutable;
+
+/**
+ * Class MyDateTime
+ *
+ * @package    Somnambulist\EntityAudit\Tests\Stubs
+ * @subpackage Somnambulist\EntityAudit\Tests\Stubs\MyDateTime
+ */
+class MyDateTime extends DateTimeImmutable
+{
+
+}